backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
langchain
celery
redis
pillow
opencv-python
paddleocr
qdrant-client
transformers
sentencepiece
numpy
python-multipart
requests
beautifulsoup4
pdfplumber
python-docx 
google-cloud-storage
google-cloud-vision
spacy
psycopg2-binary
python-dotenv
tqdm
pinecone
pydantic-settings 
cohere 